added initial code for koalagains in dodao home page
